+++ CloudShell.SecretsVaultService/SecretsVaultServiceStore.cs
@@ -1,4 +1,3 @@
-using CloudShell.Providers.Configuration;
 using Microsoft.Extensions.Options;
 using System.Text.Json;